+++ apps/extension/src/utils/host/__tests__/free-api.test.ts
@@ -1,9 +1,13 @@
-import { googleTranslate, microsoftTranslate } from '../translate/api'
+import { deeplxTranslate, googleTranslate, microsoftTranslate } from '../translate/api'
 
 describe('googleTranslate', () => {
   it('should translate text', async () => {
     const result = await googleTranslate('Library', 'en', 'zh')
     expect(result).toBe('图书馆')
+  })
+  it('should translate text to traditional chinese', async () => {
+    const result = await googleTranslate('Library', 'en', 'zh-TW')
+    expect(result).toBe('圖書館')
   })
 })
 
@@ -12,18 +16,19 @@ describe('microsoftTranslate', () => {
     const result = await microsoftTranslate('Library', 'en', 'zh')
     expect(result).toBe('图书馆')
   })
-})
-
-describe('googleTranslateZhTW', () => {
   it('should translate text to traditional chinese', async () => {
-    const result = await googleTranslate('Library', 'en', 'zh-TW')
+    const result = await microsoftTranslate('Library', 'en', 'zh-TW')
     expect(result).toBe('圖書館')
   })
 })
 
-describe('microsoftTranslateZhTW', () => {
+describe('deeplxTranslate', () => {
+  it('should translate text', async () => {
+    const result = await deeplxTranslate('Library', 'en', 'zh')
+    expect(result).toBe('图书馆')
+  })
   it('should translate text to traditional chinese', async () => {
-    const result = await microsoftTranslate('Library', 'en', 'zh-TW')
+    const result = await deeplxTranslate('Library', 'en', 'zh-TW')
     expect(result).toBe('圖書館')
   })
 })
