+++ apps/extension/src/utils/host/translate/api/google.ts
@@ -1,0 +1,60 @@
+export async function googleTranslate(
+  sourceText: string,
+  fromLang: string,
+  toLang: string,
+): Promise<string> {
+  const params = {
+    client: 'gtx',
+    sl: fromLang,
+    tl: toLang,
+    dt: 't',
+    strip: 1,
+    nonced: 1,
+    q: encodeURIComponent(sourceText),
+  }
+
+  const queryString = Object.keys(params)
+    .map(key => `${key}=${params[key as keyof typeof params]}`)
+    .join('&')
+
+  const resp = await fetch(
+    `https://translate.googleapis.com/translate_a/single?${queryString}`,
+    {
+      method: 'GET',
+    },
+  ).catch((error) => {
+    throw new Error(`Network error during translation: ${error.message}`)
+  })
+
+  if (!resp.ok) {
+    const errorText = await resp
+      .text()
+      .catch(() => 'Unable to read error response')
+    throw new Error(
+      `Translation request failed: ${resp.status} ${resp.statusText}${
+        errorText ? ` - ${errorText}` : ''
+      }`,
+    )
+  }
+
+  try {
+    const result = await resp.json()
+
+    if (!Array.isArray(result) || !Array.isArray(result[0])) {
+      throw new TypeError('Unexpected response format from translation API')
+    }
+
+    const translatedText = result[0]
+      .filter(Array.isArray)
+      .map(chunk => chunk[0])
+      .filter(Boolean)
+      .join('')
+
+    return translatedText
+  }
+  catch (error) {
+    throw new Error(
+      `Failed to parse translation response: ${(error as Error).message}`,
+    )
+  }
+}
