+++ apps/extension/src/utils/host/translate/api/microsoft.ts
@@ -1,0 +1,73 @@
+export async function microsoftTranslate(
+  sourceText: string,
+  fromLang: string,
+  toLang: string,
+): Promise<string> {
+  const effectiveFromLang = fromLang === 'auto' ? '' : fromLang
+
+  const token = await refreshMicrosoftToken()
+
+  const resp = await fetch(
+    `https://api-edge.cognitive.microsofttranslator.com/translate?from=${effectiveFromLang}&to=${toLang}&api-version=3.0&includeSentenceLength=true&textType=html`,
+    {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'Ocp-Apim-Subscription-Key': token,
+        'Authorization': `Bearer ${token}`,
+      },
+      body: JSON.stringify([{ Text: sourceText }]),
+    },
+  ).catch((error) => {
+    throw new Error(
+      `Network error during Microsoft translation: ${error.message}`,
+    )
+  })
+
+  if (!resp.ok) {
+    const errorText = await resp
+      .text()
+      .catch(() => 'Unable to read error response')
+    throw new Error(
+      `Microsoft translation request failed: ${resp.status} ${resp.statusText}${
+        errorText ? ` - ${errorText}` : ''
+      }`,
+    )
+  }
+
+  try {
+    const result = await resp.json()
+
+    if (!Array.isArray(result) || !result[0]?.translations?.[0]?.text) {
+      throw new Error(
+        'Unexpected response format from Microsoft translation API',
+      )
+    }
+
+    return result[0].translations[0].text
+  }
+  catch (error) {
+    throw new Error(
+      `Failed to parse Microsoft translation response: ${(error as Error).message}`,
+    )
+  }
+}
+
+async function refreshMicrosoftToken(): Promise<string> {
+  try {
+    const resp = await fetch('https://edge.microsoft.com/translate/auth')
+
+    if (!resp.ok) {
+      throw new Error(
+        `Failed to refresh Microsoft token: ${resp.status} ${resp.statusText}`,
+      )
+    }
+
+    return await resp.text()
+  }
+  catch (error) {
+    throw new Error(
+      `Error refreshing Microsoft token: ${(error as Error).message}`,
+    )
+  }
+}
